Organizes code and tests, Fixes CS
