+++ background-processing/appengine-frontend/app.yaml
@@ -1,4 +1,4 @@
-runtime: php72
+runtime: php73
 
 env_variables:
   APP_DEBUG: true
