+++ background-processing/cloud-run-backend/Dockerfile
@@ -5,10 +5,6 @@ WORKDIR /srv/
 # Enable PHP gRPC extension
 RUN echo "extension=grpc.so" > grpc.ini
 
-# Configure the container to serve to "router.php" instead of to "index.php"
-RUN mkdir .googleconfig \
-    && echo '{"entrypointContents": "serve router.php"}' > .googleconfig/app_start.json
-
 # Copy over all application files
 COPY . /srv/
 
@@ -21,4 +17,7 @@ ENV GAE_RUNTIME php73
 ENV FUNCTION_TARGET translateString
 
 # Start NGINX and PHP-FPM
+# NOTE: You can configure the container to serve a different endpoint by
+# creating a `.googleconfig/app_start.json` file with the contents
+# {"entrypointContents": "some entrypoint"}
 ENTRYPOINT ["/start"]
