+++ background-processing/cloud-run-backend/composer.json
@@ -3,5 +3,10 @@
     "google/cloud-translate": "^1.4",
     "google/cloud-firestore": "^1.7",
     "symfony/http-foundation": "^4.3"
+  },
+  "autoload": {
+    "files": [
+        "functions.php"
+    ]
   }
 }
