+++ background-processing/test/phpunit-backend.xml.dist
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  Copyright 2019 Google LLC
+
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+<phpunit bootstrap="../cloud-run-backend/vendor/autoload.php" convertWarningsToExceptions="false">
+    <testsuites>
+        <testsuite name="PHP Getting Started background processing backend tests">
+            <directory>backend</directory>
+        </testsuite>
+    </testsuites>
+</phpunit>
