+++ background-processing/test/phpunit-frontend.xml.dist
@@ -14,10 +14,10 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 -->
-<phpunit bootstrap="appengine-frontend/vendor/autoload.php" convertWarningsToExceptions="false">
+<phpunit bootstrap="../appengine-frontend/vendor/autoload.php" convertWarningsToExceptions="false">
     <testsuites>
         <testsuite name="PHP Getting Started background processing frontend tests">
-            <directory>test/frontend</directory>
+            <directory>frontend</directory>
         </testsuite>
     </testsuites>
     <php>
