+++ background-processing/test/phpunit-system.xml.dist
@@ -14,10 +14,13 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 -->
-<phpunit bootstrap="cloud-run-backend/vendor/autoload.php" convertWarningsToExceptions="false">
+<phpunit bootstrap="../appengine-frontend/vendor/autoload.php" convertWarningsToExceptions="false">
     <testsuites>
-        <testsuite name="PHP Getting Started background processing backend tests">
-            <directory>test/backend</directory>
+        <testsuite name="PHP Getting Started background processing system tests">
+            <directory>system</directory>
         </testsuite>
     </testsuites>
+    <php>
+        <env name="PHPUNIT_TESTS" value="1"/>
+    </php>
 </phpunit>
